Move prow pluging config sharding to pkg/prowconfigsharding
